rename component file to match component name
