+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export * from './ConvertFileTypeSelector';
+export * from './ConvertImageFileTypeSelector';
 export * from './ImageList';
 export * from './InputImageArea';
 export * from './ResultBar';
